Refactor UnitStatusParameterState
